删除，修改菜品

+++ sky-server/src/main/java/com/sky/mapper/EmployeeMapper.java
@@ -52,5 +52,5 @@ public interface EmployeeMapper {
      * @return
      */
     @Select("select * from employee where id = #{id}")
-    Employee getBtId(Integer id);
+    Employee getBtId(Long id);
 }
